Fixed JSON bug.
contacts.json (supplied for the task) was formatted differently than required. I believe the json was formated to be included into the code.  I have created contacts.json.fix which is correctly formatted to be read in as a file.
The original file when being accessed throwed the error "Unexpected token n at Object.parse (native)" this was because the name was not enclosed in any quotes, it had to be enclosed in double quotes. While the value was enclosed in only single quotes, it had to be doubled quotes.